__data__ moved to corona_hakab_model_data. for now a having a separate data module just makes things more complicated
