+++ src/corona_hakab_model/__data__.py
